requirements.txt
absl-py==1.2.0
antlr4-python3-runtime==4.8
astor==0.8.1
attrs==21.4.0
backcall==0.2.0
baselines==0.1.5
brotlipy==0.7.0
cachetools==4.2.4
captum==0.2.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.12
cilog==1.2.2
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
coverage==5.5
cryptography==35.0.0
cycler==0.11.0
Cython==0.29.24
dataclasses==0.8
decorator==5.1.1
dill==0.3.4
dive-into-graphs==0.1.2
entrypoints==0.4
et-xmlfile==1.1.0
future==0.18.2
gast==0.2.2
google-auth==2.11.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googledrivedownloader==0.4
graphviz==0.19
greenlet==1.1.1
grpcio==1.48.2
gym==0.21.0
h5py==2.10.0
hydra-core==1.1.1
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
isodate==0.6.1
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.36.0
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
more-itertools==8.12.0
mpi4py==3.1.2
mpmath==1.2.1
mypy-extensions==0.4.3
nest-asyncio==1.5.5
networkx==1.11
numba==0.53.1
numpy==1.19.5
oauthlib==3.2.1
olefile==0.46
omegaconf==2.1.1
opencv-python==4.7.0.72
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pip==21.2.2
pluggy==1.0.0
progressbar2==3.55.0
prompt-toolkit==3.0.29
protobuf==3.19.5
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.19.1
pycosat==0.6.3
pycparser==2.20
pyglet==1.5.16
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.4
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.8.2
python-utils==2.5.6
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
rdflib==5.0.0
rdkit-pypi==2021.9.3
reportlab==3.5.67
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==0.24.2
scipy==1.5.4
setuptools==58.0.4
shap==0.40.0
six==1.16.0
slicer==0.0.7
SQLAlchemy==1.4.22
sympy==1.9
tabulate==0.8.9
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow-estimator==1.15.1
termcolor==1.1.0
threadpoolctl==3.0.0
toml==0.10.2
tornado==6.1
tqdm==4.63.0
traitlets==4.3.3
typed-argument-parser==1.5.4
typing_extensions==4.1.1
typing-inspect==0.7.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.3
wheel==0.37.0
wrapt==1.13.3
yacs==0.1.8
zipp==3.6.0
zmq==0.0.0